Cleanup some unused variables

+++ input/connect/joypad_connection.c
@@ -151,7 +151,7 @@ joypad_connection_t *pad_connection_init(unsigned pads)
    if (!(joyconn = (joypad_connection_t*)calloc(pads+1, sizeof(joypad_connection_t))))
       return NULL;
 
-   for (i = 0; i < pads; i++)
+   for (i = 0; i < (int)pads; i++)
    {
       joypad_connection_t *conn  = (joypad_connection_t*)&joyconn[i];
 
@@ -240,7 +240,7 @@ void legacy_pad_connection_pad_deregister(joypad_connection_t *pad_list, pad_con
 void pad_connection_pad_deregister(joypad_connection_t *joyconn,
       pad_connection_interface_t *iface, void *pad_data)
 {
-   int i; 
+   int i;
 
    if (!iface || !iface->multi_pad)
    {
@@ -279,7 +279,7 @@ void pad_connection_pad_refresh(joypad_connection_t *joyconn,
       state = iface->status(device_data, i);
       switch(state)
       {
-         /* The pad slot is bound to a joypad 
+         /* The pad slot is bound to a joypad
             that's no longer connected */
          case PAD_CONNECT_BOUND:
             joypad = iface->joypad(device_data, i);
@@ -334,14 +334,14 @@ void pad_connection_pad_register(joypad_connection_t *joyconn,
 
    for (i = 0; i < max_pad; i++)
    {
-      int status = iface->multi_pad 
-         ? iface->status(device_data, i) 
+      int status = iface->multi_pad
+         ? iface->status(device_data, i)
          : PAD_CONNECT_READY;
       if (status == PAD_CONNECT_READY)
       {
          void *connection = NULL;
-         int found_slot   = (slot == SLOT_AUTO) 
-            ? pad_connection_find_vacant_pad(joyconn) 
+         int found_slot   = (slot == SLOT_AUTO)
+            ? pad_connection_find_vacant_pad(joyconn)
             : slot;
          if (found_slot < 0)
             continue;
@@ -383,7 +383,7 @@ int32_t pad_connection_pad_init_entry(joypad_connection_t *joyconn,
    }
    else
    {
-      /* We failed to find a matching pad. 
+      /* We failed to find a matching pad.
        * Set up one without an interface */
       RARCH_DBG("Pad was not matched. Setting up without an interface.\n");
       conn->iface      = NULL;
@@ -429,7 +429,7 @@ void pad_connection_packet(joypad_connection_t *joyconn, uint32_t pad,
 {
    if (     joyconn
          && joyconn->connected
-         && joyconn->connection 
+         && joyconn->connection
          && joyconn->iface
          && joyconn->iface->packet_handler)
       joyconn->iface->packet_handler(joyconn->connection, data, length);
